Move encrypted_import -> src/scripts src/platform/installer.
encrypted_import needs to be installed onto the image, and that's easier to
do from here. Also modified the script (and it's test) to not use common.sh,
though it now imports /usr/share/misc/shflags, which means it only runs
in the chroot or on the device.

Changes to chromeos-install-9999 are forthcoming.

BUG=chromium:275052
TEST=Ran encrypted_import_unittests.

Change-Id: Ifdbf22f1b61a3fd568f6399ea7af4bd27cf285a1
Reviewed-on: https://gerrit.chromium.org/gerrit/66449
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>