http_server: Unit tests for Server and ConnectionDelegate
Before this patch, the unit tests implemented for the server module
were actually integration tests including both the Server and the
ConnectionDelegate implementations.

This patch divides the tests in order to have independent *unit*
tests for the ConnectionDelegate and the Server implementations.
As part of this division, the Server unittest that were measuring
the bandwidth limitations implemented in ConnectionDelegate are
now unit tests of the ConnectionDelegate that use a fake clock to
avoid flakiness.

To achieve this separation, a ConnectionDelegateInterface is
added with its corresponding fake implementation.

The only remaining test using both implementations together is
P2PHttpServer.Multiple that will need to be replaced to use the
fake implementation.

BUG=chromium:269212
TEST=unit test pass!

Change-Id: Iec92613b56cc26e7559cfd29de3b687143693d5f
Reviewed-on: https://chromium-review.googlesource.com/66892
Reviewed-by: David Zeuthen <zeuthen@chromium.org>
Reviewed-by: Alejandro Deymonnaz <deymo@chromium.org>
Tested-by: Alejandro Deymonnaz <deymo@chromium.org>
Commit-Queue: Alejandro Deymonnaz <deymo@chromium.org>