shill: Service: Preserve "autoconnect=false" settings
If the AutoConnect flag is set to false via DBus, make sure that
this value does not change automatically when connected by also
setting the woefully misnamed "favorite" flag.  While here,
remove the force of |favorite_| to true on Load() as it may leave
a saved but not-yet-connected service in an non-autoconnect state.

The |favorite_| flag, is really become only a marker for whether
or not the |auto_connect_| flag should be modified on connection.

BUG=chromium:246580,chromium:219209
TEST=Unit test

Originally-reviewed-on: https://gerrit.chromium.org/gerrit/66947
(cherry picked from commit 8726d8d12d97b4b20ac7b1fd5e6046cda2ad2393)

Change-Id: Ia86ce7cdb5b895ebde5ec9010df8ba8b02fb4264
Reviewed-on: https://chromium-review.googlesource.com/167053
Tested-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>