power: Turn backlight off in docked mode.
powerd previously depended on the backlight driver turning
off the backlight when the device's lid is closed, but this
doesn't happen on all hardware. This isn't a problem when
the system is suspended, but it is when the system stays on
in docked mode.

BUG=chromium:242308
TEST=added unit tests; manually tested that backlight is
     turned off on snow when lid is mostly closed

Change-Id: I77fb3abd8c3aafd57d8b7c02d9f9ae6b1f972be0
Reviewed-on: https://chromium-review.googlesource.com/167495
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>