Remove redundant time field from ParsedEvent
An earlier change refactored struct ParsedEvent to have its own
timestamp field.  This was used to sort the events by time. However,
the patch failed to set this field when deserializing sample events
from a protobuf.

Instead of filling in the field at deserialization, just remove it
altogether.  It is only used during sorting.  So just manually extract
the timestamp data when sorting.

BUG=none
TEST=unit tests pass

Change-Id: Ibb11233cfdfd23c478cdc9c74c7f2aa207d26b89
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/167392
Reviewed-by: Ahmad Sharif <asharif@chromium.org>