http_server: Replace time-based multiple client unit test
The http_server/server unit test includes a test to check if the
server can handle multiple connections. This test had timing issues
that could make the test fail, and also was using the real
ConnectionDelegate implementation.

This fix reimplements that test to avoid the time delays and execute
the test in a synchronized way.

BUG=chromium:269212
TEST=Ran http_unittest [faster!]

Change-Id: If6b94c40039ffeb9b9dcd0465b56e96f72bf18b0
Reviewed-on: https://chromium-review.googlesource.com/167398
Reviewed-by: David Zeuthen <zeuthen@chromium.org>
Tested-by: Alejandro Deymo <deymo@chromium.org>
Commit-Queue: Alejandro Deymo <deymo@chromium.org>