Replace bool by gboolean on DBus generated calls.
The code generated for use with dbus-glib expects the boolean "b"
arguments to receive a gboolean variable and not a bool. This is
problem for both gboolean* and gboolean arguments because
sizeof(gboolean) is typically 4 while sizeof(bool) is 1. This can
led to uninitialized memory or segmentation faults.

BUG=None
TEST=update_engine_client calls still work.

Change-Id: I4036d63c875859fd73f2ff5a1585494ebe4d3603
Reviewed-on: https://chromium-review.googlesource.com/167526
Reviewed-by: David Zeuthen <zeuthen@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>