Refactor Encrypt/DecryptWithTpm into their own Crypto methods.
The ability to encrypt and decrypt with a randomly generated and sealed
key is something that we want to reuse for many other purposes. This CL
pulls out this functionality out of the Attestation class and adds a few
more unittests to test this specific functionality.

BUG=chromium:264354
TEST=Unittests.

Change-Id: Id1376400e8e714f34094531b3433744e66f00ea6
Reviewed-on: https://chromium-review.googlesource.com/167720
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>