power: Avoid shutting down when starting in docked mode.
powerd used to suspend or shut the system down immediately
if the lid was closed when powerd started. This caused
problems if the user also had an external display connected
-- if powerd restarts while in docked mode, the system would
suspend or shut down immediately.

This change makes the StateController class wait up to ten
seconds for notification from Chrome about whether the
system is in presentation mode or not. If so, the system
enters docked mode instead of going away.

Note that systems still refuse to boot in docked mode in
some cases due to vboot shutting them down if the lid is
closed.

BUG=chromium:277091
TEST=added unit tests; manually checked that system returns
     to docked mode after restarting powerd

Change-Id: I4a1c941546b9d6a8d69ed0a4bccab1f798a265c4
Reviewed-on: https://chromium-review.googlesource.com/167985
Reviewed-by: David Moore <davemoore@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>