Modify encrypted_import to have zero library dependancies.
This should allow the script to run outside the chroot, inside the chroot,
or on the device. Previously, we couldn't run outside the chroot.

It lowers the quality of the command line parsing by switching to positional
arguments only.

This change also switches to only validating .crx files. All other files in
the source directory will be ignored.

BUG=chromium:276603
TEST=Unittests (inside and outside chroot)

Change-Id: Ib4fd7a9ca8f65a5dc2d2214494237fb7fd2f08b3
Reviewed-on: https://chromium-review.googlesource.com/167921
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Dmitry Polukhin <dpolukhin@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>