netfilter-queue: Accept multicast frames
Accept multicast packets for applications that request them.
To accomplish this, when a listener is created and the destination
port of the listener is the same as its source port, add another
field to the listener called "destination".  If we receive a
multicast packet, instead of discarding it, we check if any
listeners have this destination set, then perform the normal
port and netmask checks.

While here, add a "--verbose" command line flag to display VLOG()
messages.

BUG=chromium:275737
TEST=Unit tests

Change-Id: Iee8b2501a179e63786bbebacb3bf67bad6abe842
Reviewed-on: https://chromium-review.googlesource.com/168510
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>