power: Add get_powerd_initial_backlight_level.
This adds a helper program that exercises enough powerd code
to print the default backlight brightness level.

It's intended to be used by battery-life tests that need to
stop the powerd process but also want to use the brightness
that would be set by default on a given system.

BUG=chromium:288793
TEST=manual: used brightness keys and then ran tool with AC
     plugged and unplugged

Change-Id: I1838ff907bb973aa9050c15239592f051d53ef12
Reviewed-on: https://chromium-review.googlesource.com/168843
Commit-Queue: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>