Make chaps database permissions more robust.
We have seen instances where chaps database permissions / ownership are not
right and there is no workaround short of removing the account.  Although the
root cause is not known, this CL checks all permissions and ownership before
loading a token and makes any necessary corrections, logging warnings if
corrections are necessary.  If the permissions or ownership cannot be fixed,
the token will not be loaded.

BUG=chromium:268974
TEST=unit, manual

Change-Id: Id2936753339da51ae905d044863aafa3af4ab083
Reviewed-on: https://chromium-review.googlesource.com/169750
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Commit-Queue: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>