shill: Adds ability to scan just previously seen frequencies.
We'll be trying a couple more progressive scan algorithms (the previous
ones were fine but we're experimenting) and this CL permits and tests
them.

NOTE: the word 'algorithm', below, may be overreaching.  Probably, it
should be replaced with 'set of numbers' or some other, much less
impressive, term.  Go forth with this understanding -- you've been
warned.

One algorithm does a single scan but only of previously-seen
frequencies.  This code permits WiFi to define that configuration (but
doesn't add code to set the flag).  A test is added to verify that the
mechanism works.

Another algorithm uses different min/max values (all previous groups in
the field trial set min and max to the same value).  This algorithm
will permit, when the number of previously seen frequencies is non-zero
but less than the requested maximum, scanning of just previously seen
frequencies.  The scanning code always supported this but a test was
added to verify the specific scenario this algorithm is intended to
address.

The code that enables these modes will be added in a future CL.

BUG=chromium:294484
TEST=unittest (even new ones)

Change-Id: I072325125d2b10a15bc210861501f2014f0cb1da
Reviewed-on: https://chromium-review.googlesource.com/170006
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Commit-Queue: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>