suspend_stress_test: Always enable pm_print_times
The way pm_print_times is activated by default in powerd_suspend is
really weird (and possibly unintended), and means that it never works
on kernels with serial console in the command line. We definitely always
want it when running stress tests, so the suspend_stress_test script
should override that value the way the autotests already do.

BUG=None
TEST=Manual

Change-Id: I1aec889b62644bedb2ca06489fe27d5040989622
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/169934
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>