shill: Enables new field trial groups
We'll be trying a couple more progressive scan algorithms (the previous
ones were fine but we're experimenting) and this CL adds those
algorithms as new field trial groups.

One algorithm does a single scan, only of previously-seen frequencies.
The idea is that, in nearly all cases, we'll find a good BSS in a scan
of all previously seen frequencies and that, since 75% of the users have
seen less than 6 or 7 frequencies and 50% have less than 4, 'all
frequencies' is not too large of a group in the worst case and is a
pretty small group in more then half the cases.  Note that if we don't
find a BSS in a scan, the code falls back to a complete scan.  This
algorithm is represented by two identical groups to help determine
whether the size of the field trial groups are large enough to make the
results statistically significant.

Another algorithm uses different min/max values and a group was added
with those parameters.  This allows machines that have a very small set
of previously-seen frequencies to have very short scan times, machines
that have a large set of previously-seen frequencies to have their scans
broken up to try to find a BSS without searching all of those
frequencies, and scans that don't find anything in the previously-seen
list to scan just the frequencies that haven't just been scanned.

Previous algorithms used the same min and max values -- this makes every
scan (except, possibly, the last) the same size (the min/max value).
The percentage being irrelevant means that we had 2 sets of 2 identical
groups: one with 4 frequencies per scan and one with 8.  4 frequencies
hopes that we find a BSS early while 8 frequencies assumes that we
usually won't and tries not to incur the overhead of so many scans.

BUG=chromium:294484
TEST=unittest and manual tests.  The manual test involved the following
(note, exact commands are preceded by '>' where descriptions of commands
are not):

Do the following on the DUT in a crosh tab (ctrl-alt-T):

  * > shell
  * > sudo bash
  * > rm -f /home/chronos/.progressive_scan
  * > echo "5" > /home/chronos/.progressive_scan_variation
  * In order to edit /etc/init/shill.conf, you may have to:
    > sudo mount -o remount,rw /dev/root /
  * Set shill's log level:
    - Edit /etc/init/shill.conf
    - Change the 'env SHILL_LOG_LEVEL' line to read:
          env SHILL_LOG_LEVEL=-10
    - Change the 'env SHILL_LOG_SCOPES' line to read:
          env SHILL_LOG_SCOPES=wifi
    - Save the file
  * > stop shill
  * Modify the previous frequency list:
    - Edit /var/cache/shill/default.profile
    - Search for a section marked [provider_of_wifi]
    - Remove all 'Frequencies' lines after 'Frequencies0'
    - Replace the text beyond the first semicolon with the following:
          2412:1;2437:1;2462:1;
      That line should look now like:
          Frequencies0=@xxxx;2412:1;2437:1;2462:1;
      where 'xxxx' is the same number that was there, before.
  * > echo '' > /var/log/net.log
  * > start shill
  * Wait 10 seconds
  * > cp /var/log/net.log /tmp/net.log
  * > grep 'Progressive scan (field trial 5' /tmp/net.log
    - Verify that you find at least one copy of that line.
  * > grep -A 10 'We have requested scan frequencies' /tmp/net.log
    - Verify that only the following frequencies are requested: 2412, 2437, 2462
      (if there's more than one 'We have requested' line, look at the
      first)
  * > grep 'Scanned .* frequencies (.* remaining)' /tmp/net.log
    - Verify number of scanned frequencies is 3
    - Verify number of frequencies remaining is 0
      (if there's more than one 'Scanned x frequencies' line, look at
      the first)

- on to test 2 -

  * > echo "7" > /home/chronos/.progressive_scan_variation
  * > stop shill
  * Modify the previous frequency list (exactly like before):
    - Edit /var/cache/shill/default.profile
    - Search for a section marked [provider_of_wifi]
    - Remove all 'Frequencies' lines after 'Frequencies0'
    - Replace the text beyond the first semicolon with the following:
          2412:1;2437:1;2462:1;
      That line should look now like:
          Frequencies0=@xxxx;2412:1;2437:1;2462:1;
      where 'xxxx' is the same number that was there, before.
  * > echo '' > /var/log/net.log
  * > start shill
  * > Wait 10 seconds
  * > cp /var/log/net.log /tmp/net.log
  * > grep 'Progressive scan (field trial 7' /tmp/net.log
    - Verify that you find at least one copy of that line.
  * > grep -A 10 'We have requested scan frequencies' /tmp/net.log
    - Verify that only the following frequencies are requested: 2412, 2437, 2462
      (if there's more than one 'We have requested' line, look at
      the first)
  * > grep 'Scanned .* frequencies (.* remaining)' /tmp/net.log
    - Verify that EITHER:
      o The number of scanned frequencies is 3 and remaining is NOT 0, or
      o The number of scanned frequencies is greater than 3
        (if there's more than one 'Scanned x frequencies' line, look at
        the first)

Change-Id: I191f4fb806954d0d61910e4463a1b3bc3b8a25ee
Reviewed-on: https://chromium-review.googlesource.com/170007
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>
Commit-Queue: Wade Guthrie <wdg@chromium.org>