set_cellular_ppp: move service to default profile
When a cellular dongle requires a PPP username and password,
we can find ourselves in a situation where it is impossible
to add new users to the device.

The problem is that we can only add new users at the login
screen. And to add new users, we need a network connection.
But the login screen doesn't provide any way to configure
PPP parameters, so we can't connect to a network that requires
them.

To work around this, we have the set_cellular_ppp command
move the affected Service to the default Profile. Thus, we
can add users by a) logging in as guest, b) running set_cellular_ppp,
c) logging out, and then d) adding the new user.

The change includes outputing a message to the user, informing
him or her that the connection will be available to other users.

BUG=chromium:294277
TEST=manual (see below)

Manual testing
--------------
 1. boot system, without any dongle plugged in
 2. open root shell on VT2
 3. stop shill
 4. rm /var/cache/shill/default.profile
 5. start shill
 6. log in as guest
 7. plug in dongle, wait for it to connect
 8. open chrome://network, verify that the cellular Service has
    Profile=/profile/chronos/shill, rather than /profile/default
 9. open crosh
10. set_cellular_ppp -u foo
11. observe message informing user that the connection will be
    shared
12. reload chrome://network, verify that the cellular Service has
    Profile=/profile/default
13. grep Cellular.PPP.Username /var/cache/shill/default.profile
    as root. this should show a match.
14. grep Cellular.PPP.Username /var/run/shill/user_profiles/chronos/shill.profile
    this should NOT show a match.

Change-Id: I89229a7772d872b2880b9952d05f8b4281755baf
Reviewed-on: https://chromium-review.googlesource.com/170103
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Ben Chan <benchan@chromium.org>