Move DBus service constants from flimflam namespace to shill namespace.
As the first step to deprecate the flimflam namespace, this CL moves all
DBus service constants from the flimflam namespace to the shill
namespace, and make the flimflam namespace as an alias of the shill
namespace. The namespace alias will be removed in a follow-up CL once
all code has switched to use the shill namespace.

BUG=chromium:295425
CQ-DEPEND=CL:170097
TEST=Tested the following:
1. Build system_api, shill, update_engine, and chrome.
2. Trybot runs on x86, amd64, arm platforms.

Change-Id: Ic4b979099d5765bfa7033ced1397791b41b9272c
Reviewed-on: https://chromium-review.googlesource.com/170123
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Ben Chan <benchan@chromium.org>