shill: LinkMonitor: Test reliability of unicast responses
Some gateways do not support replying to unicast ARP.  As such,
it is on tenuous grounds that we use the unicast test to trigger
a link monitor failure.  This change delays counting of unicast
failures until it is clear that unicast replies are the norm
with the currently connected gateway.

BUG=chromium:237290,chromium:295844
TEST=Unit tests

Change-Id: I3f60c09069a31ad93aa4c71fe8d5b3e0224bda34
Reviewed-on: https://chromium-review.googlesource.com/170473
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>