utils: Count metadata lines in piped perf report
The chromiumos perf tool used by quipper does not show metadata in
reports for piped perf data, even if the file contains metadata.
However, other systems' perf tools might still show the metadata.  That
is unexpected and can cause a report parsing error.

This patch handles both cases, where the metadata is shown and where it
is not shown.

BUG=none
TEST=unit tests pass

Change-Id: If5fa2886009f9304955912531a24719601ea4c22
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/170333