shill: Manager: Create and emit signals for the ConnectedState property
As opposed to the "State" property of the Manager which is strictly
"offline" or "online", this new property returns the connected state
of the highest-ranked service.  This allows observers to tell whether
or not to expect to have wider network connectivity by waiting for
this property to be "online", with the assurance that portal detection
has completed successfully.

CQ-DEPEND=CL:170701
BUG=chromium:298568
TEST=New unit test + manual: run "dbus-monitor --system" and restart
shill.  Expect to see updates to the "ConnectionState" of the manager;
initially "idle" as the manager starts:

  signal sender=:1.20 -> dest=(null destination) serial=11 path=/; interface=org.chromium.flimflam.Manager; member=PropertyChanged
   string "ConnectionState"
   variant       string "idle"

then ultimately if the connection succeeds:

signal sender=:1.20 -> dest=(null destination) serial=62 path=/; interface=org.c
hromium.flimflam.Manager; member=PropertyChanged
   string "ConnectionState"
   variant       string "online"

Change-Id: I2c34e33c6742cbe7516af7208131e76344b2f690
Reviewed-on: https://chromium-review.googlesource.com/170704
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>