Pass a temporary file to Chrome for termination messaging
The session_manager should be able to send Chrome a short message when
it decides the browser needs to die. The communication process needs to have
the following characteristics:

1. Breakpad misbehaving must not be able to block or crash the session_manager
2. Ideally, the session_manager won't block at all.
3. Breakpad should not block forever trying to read messages, either
4. Sending no message at all should be handled without any sort of timing out.

Pretty much every kind of in-memory communications channel -- fifos,
pipes, unix domain sockets -- has some kind limit on the amount of
data that can be sent before the other end begins consuming. Relying
on the size of that buffer for program correctness is a bad, bad
idea. If the session_manager just naively writes to some such
communication channel, it is possible that the session manager could
hang. This would happen if the message is larger than the kernel
buffer for the chosen communication channel and Breakpad is not
actively reading on the other side. This would be A Bad
Thing. Non-blocking IO would prevent the session manager from actually
hanging, but the code would need to be integrated into the session
manager event loop in order for it to continue processing other kinds
of events while continuing to send data -- and we'd need to give up at
some point. Such a solution could probably be made to work, but at the
cost of significant added complexity.

It's much simpler to use the tmpfs in /tmp as a huge buffer and assume
that we're never going to send a message large enough that we fill it
up. So, instead of some kind of FD that the session_manager will share
the browser, the session_manger just creates a temp file and passes
the path on the browser command line. In the event that the session
manager decides to abort the browser, it will try to write a message
to this file. The code makes a best-effort to write to tmpfs and bails
if it doesn't succeed. Breakpad can then attempt to read from this file
and include the message in its crash report.

BUG=chromium:233833
TEST=unit, suite:smoke
CQ-DEPEND=I6f5ce838ecbb6dd29793881c665ad7e398b1bd05

Change-Id: I5db5247a4cd369d4b729c4a02823e0c09e6ee78f
Reviewed-on: https://chromium-review.googlesource.com/171527
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>